// src/app/api/models/available-slot-response-dto.ts
/* tslint:disable */
 
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { SlotDetailDto } from '../models/slot-detail-dto';
export interface AvailableSlotResponseDto {

  /**
   * Fecha para la cual se listan los slots
   */
  date: string;

  /**
   * Mensaje adicional (ej. si el día está bloqueado)
   */
  message?: string;

  /**
   * Lista de slots disponibles y no disponibles para el día
   */
  slots: Array<SlotDetailDto>;
}